Move outlook specific methods to outlook util
